[WIP] Add experimental support for BitShuffle filter.
The BitShuffle filter "transposes" integer / floating point data
by, for a 8KB block, writing out the most-significant-bits of all
the data, followed by the second-most-significant, etc.  In the
case where there is repetition across the higher-order-bits (common
in floating point cases, compressed floating point, or integer-offset
sequences), this allows for improved compression.

In particular, on a CMS NanoAOD file, we see:
- 126MB for zlib-1
- 112MB for zlib-6
- 133MB for LZ4-6
- 107MB for LZ4-6 with BitShuffle

So, the BitShuffle filter - when applied to branches composed of
primitive types - results in a significant improvement for LZ4 (where
it is even better than LZ4).

We theorize that this is because LZ4 is composed of a relatively simple
LZ77-based pass, which looks for common patterns in the uncompressed
buffer, while other algorithms (ZLIB/DEFLATE, LZMA) do a LZ77 pass
plus some sort of Huffman coding (which replaces common symbols with
prefix codes) or similar (LZMA has an interesting pass of integer
range-based encoding).  Hence, we are using special properties of
primitive types to pre-condition the branch data.  See the following
paper for a visual example of how this works:

https://arxiv.org/pdf/1503.00638.pdf

Interestingly enough, it appears that BitShuffle causes a small
increase in size for LZMA (as LZMA tries to parse things as integers
directly).

This code is a proof-of-concept -- just enough to write out encoded
files.  It does not pay attention to speed or memory costs.